replace old randlanet
