--- ml3d/configs/randlanet_new_parislille3d.yml
@@ -1,61 +1,0 @@
-dataset:
-  name: ParisLille3D
-  dataset_path: # path/to/your/dataset
-  cache_dir: ./logs/cache
-  class_weights: [53033221, 25821560,   546190, 54417,
-        180638,  3998129, 160899, 3441844, 9681606]
-  ignored_label_inds:
-  - 0
-  num_points: 65536
-  test_result_folder: ./test
-  val_files:
-  - Lille2.ply
-  use_cache: true
-  steps_per_epoch_train: 50
-  steps_per_epoch_valid: 10
-  sampler:
-    name: SemSegSpatiallyRegularSampler
-model:
-  name: RandLANet
-  batcher: DefaultBatcher
-  ckpt_path: # path/to/your/checkpoint
-  num_neighbors: 16
-  num_layers: 5
-  num_points: 65536
-  num_classes: 9
-  ignored_label_inds: [0]
-  sub_sampling_ratio: [4, 4, 4, 4, 2]
-  in_channels: 3
-  dim_features: 8
-  dim_output: [16, 64, 128, 256, 512]
-  grid_size: 0.06
-  weight_decay: 0.001
-  augment:
-    recenter:
-      dim: [0, 1, 2]
-    rotate:
-      method: vertical
-    scale:
-      min_s: 0.9
-      max_s: 1.1
-    noise:
-      noise_std: 0.01
-    RandomHorizontalFlip:
-      axes: [0, 1]
-pipeline:
-  name: SemanticSegmentation
-  adam_lr: 0.01
-  batch_size: 2
-  learning_rate: 0.01
-  main_log_dir: ./logs
-  max_epoch: 100
-  save_ckpt_freq: 5
-  scheduler_gamma: 0.9886
-  test_batch_size: 1
-  train_sum_dir: train_log
-  val_batch_size: 2
-  summary:
-    record_for: []
-    max_pts:
-    use_reference: false
-    max_outputs: 1
