--- ml3d/configs/randlanet_new_s3dis.yml
@@ -1,63 +1,0 @@
-dataset:
-  name: S3DIS
-  dataset_path: # path/to/your/dataset
-  cache_dir: ./logs/cache
-  class_weights: [3370714, 2856755, 4919229, 318158, 375640,
-478001, 974733, 650464, 791496, 88727, 1284130, 229758, 2272837]
-  ignored_label_inds: []
-  num_points: 40960
-  test_area_idx: 3
-  test_result_folder: ./test
-  use_cache: False
-model:
-  name: RandLANet
-  batcher: DefaultBatcher
-  ckpt_path: # path/to/your/checkpoint
-  num_neighbors: 16
-  num_layers: 5
-  num_points: 40960
-  num_classes: 13
-  ignored_label_inds: []
-  sub_sampling_ratio: [4, 4, 4, 4, 2]
-  in_channels: 6
-  dim_features: 8
-  dim_output: [16, 64, 128, 256, 512]
-  grid_size: 0.04
-  augment:
-    recenter:
-      dim: [0, 1, 2]
-    rotate:
-      method: vertical
-    scale:
-      min_s: 0.9
-      max_s: 1.1
-    noise:
-      noise_std: 0.01
-    RandomDropout:
-      dropout_ratio: 0.1
-    RandomHorizontalFlip:
-      axes: [0, 1]
-    ChromaticAutoContrast:
-      randomize_blend_factor: True
-      blend_factor: 0.5
-    ChromaticTranslation:
-      trans_range_ratio: 0.1
-    ChromaticJitter:
-      std: 0.01
-pipeline:
-  name: SemanticSegmentation
-  adam_lr: 0.01
-  batch_size: 2
-  learning_rate: 0.01
-  main_log_dir: ./logs
-  max_epoch: 100
-  save_ckpt_freq: 20
-  scheduler_gamma: 0.95
-  test_batch_size: 3
-  train_sum_dir: train_log
-  val_batch_size: 2
-  summary:
-    record_for: []
-    max_pts:
-    use_reference: false
-    max_outputs: 1
