--- ml3d/configs/randlanet_new_toronto3d.yml
@@ -1,73 +1,0 @@
-dataset:
-  name: Toronto3D
-  cache_dir: ./logs/cache
-  dataset_path: # path/to/your/dataset
-  class_weights: [41697357, 1745448, 6572572, 19136493, 674897, 897825, 4634634, 374721]
-  ignored_label_inds:
-  - 0
-  num_classes: 8
-  num_points: 65536
-  test_files:
-  - L002.ply
-  test_result_folder: ./test
-  train_files:
-  - L001.ply
-  - L003.ply
-  - L004.ply
-  use_cache: true
-  val_files:
-  - L002.ply
-  steps_per_epoch_train: 100
-  steps_per_epoch_valid: 10
-model:
-  name: RandLANet
-  batcher: DefaultBatcher
-  ckpt_path: # path/to/your/checkpoint
-  num_neighbors: 16
-  num_layers: 5
-  num_points: 65536
-  num_classes: 8
-  ignored_label_inds: [0]
-  sub_sampling_ratio: [4, 4, 4, 4, 2]
-  in_channels: 6
-  dim_features: 8
-  dim_output: [16, 64, 128, 256, 512]
-  grid_size: 0.05
-  augment:
-    recenter:
-      dim: [0, 1, 2]
-    rotate:
-      method: vertical
-    scale:
-      min_s: 0.9
-      max_s: 1.1
-    noise:
-      noise_std: 0.01
-    RandomDropout:
-      dropout_ratio: 0.1
-    RandomHorizontalFlip:
-      axes: [0, 1]
-    ChromaticAutoContrast:
-      randomize_blend_factor: True
-      blend_factor: 0.5
-    ChromaticTranslation:
-      trans_range_ratio: 0.1
-    ChromaticJitter:
-      std: 0.01
-pipeline:
-  name: SemanticSegmentation
-  adam_lr: 0.01
-  batch_size: 2
-  learning_rate: 0.01
-  main_log_dir: ./logs
-  max_epoch: 100
-  save_ckpt_freq: 5
-  scheduler_gamma: 0.9886
-  test_batch_size: 1
-  train_sum_dir: train_log
-  val_batch_size: 2
-  summary:
-    record_for: []
-    max_pts:
-    use_reference: false
-    max_outputs: 1
