+++ ml3d/configs/randlanet_s3dis.yml
@@ -13,31 +13,37 @@ model:
   name: RandLANet
   batcher: DefaultBatcher
   ckpt_path: # path/to/your/checkpoint
-  dim_feature: 8
-  dim_input: 6
-  dim_output:
-  - 16
-  - 64
-  - 128
-  - 256
-  - 512
-  grid_size: 0.04
-  ignored_label_inds: []
-  k_n: 16
-  num_classes: 13
+  num_neighbors: 16
   num_layers: 5
   num_points: 40960
-  sub_sampling_ratio:
-  - 4
-  - 4
-  - 4
-  - 4
-  - 2
-  t_normalize:
-    method: linear
-    normalize_points: False
-    feat_bias: 0
-    feat_scale: 1
+  num_classes: 13
+  ignored_label_inds: []
+  sub_sampling_ratio: [4, 4, 4, 4, 2]
+  in_channels: 6
+  dim_features: 8
+  dim_output: [16, 64, 128, 256, 512]
+  grid_size: 0.04
+  augment:
+    recenter:
+      dim: [0, 1, 2]
+    rotate:
+      method: vertical
+    scale:
+      min_s: 0.9
+      max_s: 1.1
+    noise:
+      noise_std: 0.01
+    RandomDropout:
+      dropout_ratio: 0.1
+    RandomHorizontalFlip:
+      axes: [0, 1]
+    ChromaticAutoContrast:
+      randomize_blend_factor: True
+      blend_factor: 0.5
+    ChromaticTranslation:
+      trans_range_ratio: 0.1
+    ChromaticJitter:
+      std: 0.01
 pipeline:
   name: SemanticSegmentation
   adam_lr: 0.01
