+++ ml3d/configs/randlanet_semantic3d.yml
@@ -3,8 +3,7 @@ dataset:
   dataset_path: # path/to/your/dataset
   cache_dir: ./logs/cache_small3d/
   class_weights: [5181602, 5012952, 6830086, 1311528, 10476365, 946982, 334860, 269353]
-  ignored_label_inds:
-  - 0
+  ignored_label_inds: [0]
   num_points: 65536
   test_result_folder: ./test
   use_cache: true
@@ -17,39 +16,37 @@ model:
   name: RandLANet
   batcher: DefaultBatcher
   ckpt_path: # path/to/your/checkpoint
-  dim_feature: 8
-  dim_input: 6
-  dim_output:
-  - 16
-  - 64
-  - 128
-  - 256
-  - 512
-  grid_size: 0.06
-  ignored_label_inds:
-  - 0
-  k_n: 16
-  num_classes: 8
+  num_neighbors: 16
   num_layers: 5
   num_points: 65536
-  sub_sampling_ratio:
-  - 4
-  - 4
-  - 4
-  - 4
-  - 2
-  t_normalize:
-    method: linear
-    feat_bias: 0
-    feat_scale: 255
-  t_augment:
-    turn_on: true
-    rotation_method: vertical
-    scale_anisotropic: true
-    symmetries: [true, false, false]
-    noise_level: 0.001
-    min_s: 0.8
-    max_s: 1.2
+  num_classes: 8
+  ignored_label_inds: [0]
+  sub_sampling_ratio: [4, 4, 4, 4, 2]
+  in_channels: 6
+  dim_features: 8
+  dim_output: [16, 64, 128, 256, 512]
+  grid_size: 0.05
+  augment:
+    recenter:
+      dim: [0, 1, 2]
+    rotate:
+      method: vertical
+    scale:
+      min_s: 0.9
+      max_s: 1.1
+    noise:
+      noise_std: 0.01
+    RandomDropout:
+      dropout_ratio: 0.1
+    RandomHorizontalFlip:
+      axes: [0, 1]
+    ChromaticAutoContrast:
+      randomize_blend_factor: True
+      blend_factor: 0.5
+    ChromaticTranslation:
+      trans_range_ratio: 0.1
+    ChromaticJitter:
+      std: 0.01
 pipeline:
   name: SemanticSegmentation
   adam_lr: 0.01
@@ -58,11 +55,10 @@ pipeline:
   main_log_dir: ./logs
   max_epoch: 100
   save_ckpt_freq: 5
-  scheduler_gamma: 0.95
+  scheduler_gamma: 0.9886
   test_batch_size: 1
   train_sum_dir: train_log
-  val_batch_size: 1
-  test_compute_metric: false
+  val_batch_size: 2
   summary:
     record_for: []
     max_pts:
