+++ ml3d/tf/models/randlanet.py
@@ -1,305 +1,367 @@
-# ----------------------------------------------------------------------------
-# -              Creative Commons BY-NC-SA 4.0 International                 -
-# ----------------------------------------------------------------------------
-
-# By exercising the Licensed Rights (defined below), You accept and agree to
-# be bound by the terms and conditions of this Creative
-# Commons Attribution-NonCommercial-ShareAlike 4.0 International Public
-# License ("Public License"). To the extent this Public License may be
-# interpreted as a contract, You are granted the Licensed Rights in
-# consideration of Your acceptance of these terms and conditions, and the
-# Licensor grants You such rights in consideration of benefits the Licensor
-# receives from making the Licensed Material available under these terms and
-# conditions.
-#
-# You can read the full license here
-# https://creativecommons.org/licenses/by-nc-sa/4.0/legalcode
-# ----------------------------------------------------------------------------
-
 import tensorflow as tf
 import numpy as np
 
 from tqdm import tqdm
 from sklearn.neighbors import KDTree
 
-# use relative import for being compatible with Open3d main repo
 from .base_model import BaseModel
 from ..utils import helper_tf
+from ...datasets.augment import SemsegAugmentation
 from ...utils import MODEL
 from ...datasets.utils import (DataProcessing, trans_normalize, trans_augment,
                                trans_crop_pc)
+
+
+class SharedMLP(tf.keras.layers.Layer):
+
+    def __init__(self,
+                 in_channels,
+                 out_channels,
+                 kernel_size=1,
+                 stride=1,
+                 transpose=False,
+                 bn=True,
+                 activation_fn=None):
+        super(SharedMLP, self).__init__()
+
+        if transpose:
+            self.conv = tf.keras.layers.Conv2DTranspose(filters=out_channels,
+                                                        kernel_size=kernel_size,
+                                                        strides=stride,
+                                                        padding='valid')
+        else:
+            self.conv = tf.keras.layers.Conv2D(filters=out_channels,
+                                               kernel_size=kernel_size,
+                                               strides=stride,
+                                               padding='valid')
+
+        self.batch_norm = tf.keras.layers.BatchNormalization(
+            axis=1, momentum=0.99, epsilon=1e-6) if bn else None
+        self.activation_fn = activation_fn
+
+    def call(self, input, training):
+        """Forward pass of the Module.
+
+        Args:
+            input: torch.Tensor of shape (B, dim_in, N, K)
+
+        Returns:
+            torch.Tensor, shape (B, dim_out, N, K)
+        """
+        x = self.conv(input)
+        if self.batch_norm:
+            x = self.batch_norm(x, training=training)
+        if self.activation_fn:
+            x = self.activation_fn(x)
+        return x
+
+
+class LocalSpatialEncoding(tf.keras.layers.Layer):
+
+    def __init__(self, dim_in, dim_out, num_neighbors, encode_pos=False):
+        super(LocalSpatialEncoding, self).__init__()
+
+        self.num_neighbors = num_neighbors
+        self.mlp = SharedMLP(dim_in,
+                             dim_out,
+                             activation_fn=tf.keras.layers.LeakyReLU(0.2))
+        self.encode_pos = encode_pos
+
+        # self.device = device
+
+    def gather_neighbor(self, coords, neighbor_indices):
+        """Gather features based on neighbor indices.
+
+        Args:
+            coords: torch.Tensor of shape (B, N, d)
+            neighbor_indices: torch.Tensor of shape (B, N, K)
+        
+        Returns:
+            gathered neighbors of shape (B, dim, N, K)
+
+        """
+        B, N, K = neighbor_indices.shape
+        dim = coords.shape[2]
+
+        extended_indices = tf.reshape(neighbor_indices, shape=(-1, N * K))
+        neighbor_coords = tf.gather(coords,
+                                    extended_indices,
+                                    axis=1,
+                                    batch_dims=1)
+
+        return tf.reshape(neighbor_coords, [-1, N, K, dim])
+
+    def call(self,
+             coords,
+             features,
+             neighbor_indices,
+             training,
+             relative_features=None):
+        """Forward pass of the Module.
+
+        Args:
+            coords: coordinates of the pointcloud
+                torch.Tensor of shape (B, N, 3)
+            features: features of the pointcloud.
+                torch.Tensor of shape (B, d, N, 1)
+            neighbor_indices: indices of k neighbours.
+                torch.Tensor of shape (B, N, K)
+
+        Returns:
+            torch.Tensor of shape (B, 2*d, N, K)
+        """
+        # finding neighboring points
+        B, N, K = neighbor_indices.shape
+
+        if self.encode_pos:
+            neighbor_coords = self.gather_neighbor(coords, neighbor_indices)
+
+            extended_coords = tf.tile(tf.expand_dims(
+                coords, axis=2), [1, 1, tf.shape(neighbor_indices)[-1], 1])
+            relative_pos = extended_coords - neighbor_coords
+            relative_dist = tf.sqrt(
+                tf.reduce_sum(tf.square(relative_pos), axis=-1, keepdims=True))
+            relative_features = tf.concat(
+                [relative_dist, relative_pos, extended_coords, neighbor_coords],
+                axis=-1)
+        else:
+            if relative_features is None:
+                raise ValueError(
+                    "LocalSpatialEncoding: Require relative_features for second pass."
+                )
+
+        relative_features = self.mlp(relative_features, training=training)
+
+        neighbor_features = self.gather_neighbor(tf.squeeze(features, axis=2),
+                                                 neighbor_indices)
+
+        return tf.concat([neighbor_features, relative_features],
+                         axis=-1), relative_features
+
+
+class AttentivePooling(tf.keras.layers.Layer):
+
+    def __init__(self, in_channels, out_channels):
+        super(AttentivePooling, self).__init__()
+
+        self.score_fn = tf.keras.models.Sequential(
+            ((tf.keras.layers.Dense(in_channels),
+              tf.keras.layers.Softmax(axis=-2))))
+
+        self.mlp = SharedMLP(in_channels,
+                             out_channels,
+                             activation_fn=tf.keras.layers.LeakyReLU(0.2))
+
+    def call(self, x, training):
+        """Forward pass of the Module.
+
+        Args:
+            x: tf.Tensor of shape (B, N, K, dim_in).
+
+        Returns:
+            torch.Tensor of shape (B, d_out, N, 1).
+        """
+        # computing attention scores
+        scores = self.score_fn(x)
+
+        # sum over the neighbors
+        features = tf.reduce_sum(scores * x, axis=-2,
+                                 keepdims=True)  # shape (B, d_in, N, 1)
+
+        return self.mlp(features, training=training)
+
+
+class LocalFeatureAggregation(tf.keras.layers.Layer):
+
+    def __init__(self, d_in, d_out, num_neighbors):
+        super(LocalFeatureAggregation, self).__init__()
+
+        self.num_neighbors = num_neighbors
+
+        self.mlp1 = SharedMLP(d_in,
+                              d_out // 2,
+                              activation_fn=tf.keras.layers.LeakyReLU(0.2))
+        self.mlp2 = SharedMLP(d_out, 2 * d_out)
+        self.shortcut = SharedMLP(d_in, 2 * d_out)
+
+        self.lse1 = LocalSpatialEncoding(10,
+                                         d_out // 2,
+                                         num_neighbors,
+                                         encode_pos=True)
+        self.lse2 = LocalSpatialEncoding(d_out // 2, d_out // 2, num_neighbors)
+
+        self.pool1 = AttentivePooling(d_out, d_out // 2)
+        self.pool2 = AttentivePooling(d_out, d_out)
+
+        self.lrelu = tf.keras.layers.LeakyReLU(0.2)
+
+    def call(self, coords, feat, neighbor_indices, training):
+        """Forward pass of the Module.
+
+        Args:
+            coords: coordinates of the pointcloud
+                torch.Tensor of shape (B, N, 3).
+            feat: features of the pointcloud.
+                torch.Tensor of shape (B, d, N, 1)
+            neighbor_indices: Indices of neighbors.
+
+        Returns:
+            torch.Tensor of shape (B, 2*d_out, N, 1).
+
+        """
+        # knn_output = knn(coords.cpu().contiguous(), coords.cpu().contiguous(), self.num_neighbors)
+
+        x = self.mlp1(feat, training=training)
+
+        x, neighbor_features = self.lse1(coords,
+                                         x,
+                                         neighbor_indices,
+                                         training=training)
+        x = self.pool1(x, training=training)
+
+        x, _ = self.lse2(coords,
+                         x,
+                         neighbor_indices,
+                         relative_features=neighbor_features,
+                         training=training)
+        x = self.pool2(x, training=training)
+
+        return self.lrelu(
+            self.mlp2(x, training=training) +
+            self.shortcut(feat, training=training))
 
 
 class RandLANet(BaseModel):
     """Class defining RandLANet, a Semantic Segmentation model.
     Based on the architecture
     https://arxiv.org/abs/1911.11236#
-
-    Reference Implementation - https://github.com/QingyongHu/RandLA-Net
     """
 
     def __init__(
             self,
             name='RandLANet',
-            k_n=16,  # KNN,
-            num_layers=4,  # Number of layers
-            num_points=4096 * 11,  # Number of input points
-            num_classes=19,  # Number of valid classes
+            num_neighbors=16,
+            num_layers=4,
+            num_points=4096 * 11,
+            num_classes=19,
             ignored_label_inds=[0],
             sub_sampling_ratio=[4, 4, 4, 4],
-            dim_input=3,
-            dim_feature=8,
+            in_channels=3,  # 3 + feature_dimension.
+            dim_features=8,
             dim_output=[16, 64, 128, 256],
             grid_size=0.06,
             batcher='DefaultBatcher',
             ckpt_path=None,
+            augment=None,
             **kwargs):
 
         super().__init__(name=name,
-                         k_n=k_n,
+                         num_neighbors=num_neighbors,
                          num_layers=num_layers,
                          num_points=num_points,
                          num_classes=num_classes,
                          ignored_label_inds=ignored_label_inds,
                          sub_sampling_ratio=sub_sampling_ratio,
-                         dim_input=dim_input,
-                         dim_feature=dim_feature,
+                         in_channels=in_channels,
+                         dim_features=dim_features,
                          dim_output=dim_output,
                          grid_size=grid_size,
                          batcher=batcher,
                          ckpt_path=ckpt_path,
+                         augment=augment,
                          **kwargs)
 
         cfg = self.cfg
+        self.augmenter = SemsegAugmentation(cfg.augment, seed=self.rng)
 
-        dim_feature = cfg.dim_feature
-        self.fc0 = tf.keras.layers.Dense(dim_feature, activation=None)
-        self.batch_normalization = tf.keras.layers.BatchNormalization(
-            -1, momentum=0.99, epsilon=1e-6)
-        self.leaky_relu0 = tf.keras.layers.LeakyReLU()
-
-        # ###########################Encoder############################
-        d_encoder_list = []
+        self.fc0 = tf.keras.layers.Dense(cfg.dim_features)
+        self.bn0 = tf.keras.layers.BatchNormalization(-1,
+                                                      momentum=0.99,
+                                                      epsilon=1e-6)
+        self.lr0 = tf.keras.layers.LeakyReLU(0.2)
 
         # Encoder
+        self.encoder = []
+        encoder_dim_list = []
+        dim_feature = cfg.dim_features
         for i in range(cfg.num_layers):
-            name = 'Encoder_layer_' + str(i)
-            self.init_dilated_res_block(dim_feature, cfg.dim_output[i], name)
-            dim_feature = cfg.dim_output[i] * 2
+            self.encoder.append(
+                LocalFeatureAggregation(dim_feature, cfg.dim_output[i],
+                                        cfg.num_neighbors))
+            dim_feature = 2 * cfg.dim_output[i]
             if i == 0:
-                d_encoder_list.append(dim_feature)
-            d_encoder_list.append(dim_feature)
+                encoder_dim_list.append(dim_feature)
+            encoder_dim_list.append(dim_feature)
 
-        feature = helper_tf.conv2d(True, dim_feature)
-        setattr(self, 'decoder_0', feature)
+        # self.encoder = tf.keras.models.Sequential(self.encoder)
+
+        self.mlp = SharedMLP(dim_feature,
+                             dim_feature,
+                             activation_fn=tf.keras.layers.LeakyReLU(0.2))
 
         # Decoder
-        for j in range(cfg.num_layers):
-            name = 'Decoder_layer_' + str(j)
-            dim_input = d_encoder_list[-j - 2] + dim_feature
-            dim_output = d_encoder_list[-j - 2]
+        self.decoder = []
+        for i in range(cfg.num_layers):
+            self.decoder.append(
+                SharedMLP(encoder_dim_list[-i - 2] + dim_feature,
+                          encoder_dim_list[-i - 2],
+                          transpose=True,
+                          activation_fn=tf.keras.layers.LeakyReLU(0.2)))
+            dim_feature = encoder_dim_list[-i - 2]
 
-            f_decoder_i = helper_tf.conv2d_transpose(True, dim_output)
-            setattr(self, name, f_decoder_i)
-            dim_feature = d_encoder_list[-j - 2]
+        # self.decoder = tf.keras.models.Sequential(self.decoder)
 
-        f_layer_fc1 = helper_tf.conv2d(True, 64)
-        setattr(self, 'fc1', f_layer_fc1)
-
-        f_layer_fc2 = helper_tf.conv2d(True, 32)
-        setattr(self, 'fc2', f_layer_fc2)
-
-        f_dropout = tf.keras.layers.Dropout(0.5)
-        setattr(self, 'dropout1', f_dropout)
-
-        f_layer_fc3 = helper_tf.conv2d(False, cfg.num_classes, activation=False)
-        setattr(self, 'fc', f_layer_fc3)
-
-    def init_att_pooling(self, d, dim_output, name):
-        att_activation = tf.keras.layers.Dense(d, activation=None)
-        setattr(self, name + 'fc', att_activation)
-
-        f_agg = helper_tf.conv2d(True, dim_output)
-        setattr(self, name + 'mlp', f_agg)
-
-    def init_building_block(self, dim_input, dim_output, name):
-        f_pc = helper_tf.conv2d(True, dim_input)
-
-        setattr(self, name + 'mlp1', f_pc)
-
-        self.init_att_pooling(dim_input * 2, dim_output // 2,
-                              name + 'att_pooling_1')
-
-        f_xyz = helper_tf.conv2d(True, dim_output // 2)
-        setattr(self, name + 'mlp2', f_xyz)
-
-        self.init_att_pooling(dim_input * 2, dim_output, name + 'att_pooling_2')
-
-    def init_dilated_res_block(self, dim_input, dim_output, name):
-        f_pc = helper_tf.conv2d(True, dim_output // 2)
-        setattr(self, name + 'mlp1', f_pc)
-
-        self.init_building_block(dim_output // 2, dim_output, name + 'LFA')
-
-        f_pc = helper_tf.conv2d(True, dim_output * 2, activation=False)
-        setattr(self, name + 'mlp2', f_pc)
-
-        shortcut = helper_tf.conv2d(True, dim_output * 2, activation=False)
-        setattr(self, name + 'shortcut', shortcut)
-
-    def forward_gather_neighbour(self, pc, neighbor_idx):
-        # pc:           BxNxd
-        # neighbor_idx: BxNxK
-        B, N, K = neighbor_idx.shape
-        d = pc.shape[2]
-
-        index_input = tf.reshape(neighbor_idx, shape=[-1, N * K])
-
-        features = tf.gather(pc, index_input, axis=1, batch_dims=1)
-
-        features = tf.reshape(features, [-1, N, K, d])
-
-        return features
-
-    def forward_att_pooling(self, feature_set, name):
-        # feature_set: BxNxKxd
-        num_points = feature_set.shape[1]
-        num_neigh = feature_set.shape[2]
-        d = feature_set.shape[3]
-
-        f_reshaped = tf.reshape(feature_set, shape=[-1, num_neigh, d])
-
-        m_dense = getattr(self, name + 'fc')
-        att_activation = m_dense(f_reshaped)
-        att_scores = tf.nn.softmax(att_activation, axis=1)
-
-        # print("att_scores = ", att_scores.shape)
-        f_agg = f_reshaped * att_scores
-        f_agg = tf.reduce_sum(f_agg, axis=1)
-        f_agg = tf.reshape(f_agg, [-1, num_points, 1, d])
-
-        m_conv2d = getattr(self, name + 'mlp')
-        f_agg = m_conv2d(f_agg, training=self.training)
-
-        return f_agg
-
-    def forward_relative_pos_encoding(self, xyz, neigh_idx):
-        neighbor_xyz = self.forward_gather_neighbour(xyz, neigh_idx)
-
-        xyz_tile = tf.tile(tf.expand_dims(xyz, axis=2),
-                           [1, 1, tf.shape(neigh_idx)[-1], 1])
-        relative_xyz = xyz_tile - neighbor_xyz
-        relative_dis = tf.sqrt(
-            tf.reduce_sum(tf.square(relative_xyz), axis=-1, keepdims=True))
-        relative_feature = tf.concat(
-            [relative_dis, relative_xyz, xyz_tile, neighbor_xyz], axis=-1)
-        return relative_feature
-
-    def forward_building_block(self, xyz, feature, neigh_idx, name):
-        f_xyz = self.forward_relative_pos_encoding(xyz, neigh_idx)
-        m_conv2d = getattr(self, name + 'mlp1')
-
-        f_xyz = m_conv2d(f_xyz, training=self.training)
-
-        f_neighbours = self.forward_gather_neighbour(
-            tf.squeeze(feature, axis=2), neigh_idx)
-        f_concat = tf.concat([f_neighbours, f_xyz], axis=-1)
-
-        f_pc_agg = self.forward_att_pooling(f_concat, name + 'att_pooling_1')
-
-        m_conv2d = getattr(self, name + 'mlp2')
-        f_xyz = m_conv2d(f_xyz, training=self.training)
-
-        f_neighbours = self.forward_gather_neighbour(
-            tf.squeeze(f_pc_agg, axis=2), neigh_idx)
-        f_concat = tf.concat([f_neighbours, f_xyz], axis=-1)
-        f_pc_agg = self.forward_att_pooling(f_concat, name + 'att_pooling_2')
-
-        return f_pc_agg
-
-    def forward_dilated_res_block(self, feature, xyz, neigh_idx, dim_output,
-                                  name):
-        m_conv2d = getattr(self, name + 'mlp1')
-        f_pc = m_conv2d(feature, training=self.training)
-
-        f_pc = self.forward_building_block(xyz, f_pc, neigh_idx, name + 'LFA')
-
-        m_conv2d = getattr(self, name + 'mlp2')
-        f_pc = m_conv2d(f_pc, training=self.training)
-
-        m_conv2d = getattr(self, name + 'shortcut')
-        shortcut = m_conv2d(feature, training=self.training)
-
-        result = tf.nn.leaky_relu(f_pc + shortcut)
-        return result
+        self.fc1 = tf.keras.models.Sequential(
+            (SharedMLP(dim_feature,
+                       64,
+                       activation_fn=tf.keras.layers.LeakyReLU(0.2)),
+             SharedMLP(64, 32, activation_fn=tf.keras.layers.LeakyReLU(0.2)),
+             tf.keras.layers.Dropout(0.5),
+             SharedMLP(32, cfg.num_classes, bn=False)))
 
     def call(self, inputs, training=True):
         self.training = training
-        num_layers = self.cfg.num_layers
-        xyz = inputs[:num_layers]
-        neigh_idx = inputs[num_layers:2 * num_layers]
-        sub_idx = inputs[2 * num_layers:3 * num_layers]
-        interp_idx = inputs[3 * num_layers:4 * num_layers]
-        feature = inputs[4 * num_layers]
+        cfg = self.cfg
+        num_layers = cfg.num_layers
+        coords_list = inputs[:num_layers]
+        neighbor_indices_list = inputs[num_layers:2 * num_layers]
+        subsample_indices_list = inputs[2 * num_layers:3 * num_layers]
+        interpolation_indices_list = inputs[3 * num_layers:4 * num_layers]
+        feat = inputs[4 * num_layers]
 
-        m_dense = getattr(self, 'fc0')
-        feature = m_dense(feature, training=self.training)
+        feat = self.fc0(feat)  # (B, N, dim_feature)
+        feat = self.bn0(feat, training=training)  # (B, N, dim_feature)
+        feat = self.lr0(feat)
+        feat = tf.expand_dims(feat, axis=2)  # (B, N, 1, dim_feature)
 
-        m_bn = getattr(self, 'batch_normalization')
-        feature = m_bn(feature, training=self.training)
-
-        feature = tf.nn.leaky_relu(feature)
-        feature = tf.expand_dims(feature, axis=2)
-
-        # B N 1 d
         # Encoder
-        f_encoder_list = []
-        for i in range(self.cfg.num_layers):
-            name = 'Encoder_layer_' + str(i)
-            f_encoder_i = self.forward_dilated_res_block(
-                feature, xyz[i], neigh_idx[i], self.cfg.dim_output[i], name)
-            f_sampled_i = self.random_sample(f_encoder_i, sub_idx[i])
-            feature = f_sampled_i
+        encoder_feat_list = []
+        for i in range(cfg.num_layers):
+            feat_encoder_i = self.encoder[i](coords_list[i],
+                                             feat,
+                                             neighbor_indices_list[i],
+                                             training=training)
+            feat_sampled_i = self.random_sample(feat_encoder_i,
+                                                subsample_indices_list[i])
             if i == 0:
-                f_encoder_list.append(f_encoder_i)
-            f_encoder_list.append(f_sampled_i)
+                encoder_feat_list.append(tf.identity(feat_encoder_i))
+            encoder_feat_list.append(tf.identity(feat_sampled_i))
+            feat = feat_sampled_i
 
-        m_conv2d = getattr(self, 'decoder_0')
-        feature = m_conv2d(f_encoder_list[-1], training=self.training)
+        feat = self.mlp(feat, training=training)
 
         # Decoder
-        f_decoder_list = []
-        for j in range(self.cfg.num_layers):
-            f_interp_i = self.nearest_interpolation(feature, interp_idx[-j - 1])
-            name = 'Decoder_layer_' + str(j)
+        for i in range(cfg.num_layers):
+            feat_interpolation_i = self.nearest_interpolation(
+                feat, interpolation_indices_list[-i - 1])
+            feat_decoder_i = tf.concat(
+                [encoder_feat_list[-i - 2], feat_interpolation_i], axis=-1)
+            feat_decoder_i = self.decoder[i](feat_decoder_i, training=training)
+            feat = feat_decoder_i
 
-            m_transposeconv2d = getattr(self, name)
-            concat_feature = tf.concat([f_encoder_list[-j - 2], f_interp_i],
-                                       axis=3)
-            f_decoder_i = m_transposeconv2d(concat_feature,
-                                            training=self.training)
+        scores = self.fc1(feat, training=training)
 
-            feature = f_decoder_i
-            f_decoder_list.append(f_decoder_i)
-
-        m_conv2d = getattr(self, 'fc1')
-        f_layer_fc1 = m_conv2d(f_decoder_list[-1], training=self.training)
-
-        m_conv2d = getattr(self, 'fc2')
-        f_layer_fc2 = m_conv2d(f_layer_fc1, training=self.training)
-
-        self.test_hidden = f_layer_fc2
-
-        m_dropout = getattr(self, 'dropout1')
-        f_layer_drop = m_dropout(f_layer_fc2, training=self.training)
-
-        m_conv2d = getattr(self, 'fc')
-        f_layer_fc3 = m_conv2d(f_layer_drop, training=self.training)
-
-        f_out = tf.squeeze(f_layer_fc3, [2])
-        # f_out = tf.nn.softmax(f_out)
-
-        return f_out
+        return tf.squeeze(scores, [2])
 
     def get_optimizer(self, cfg_pipeline):
         lr_schedule = tf.keras.optimizers.schedules.ExponentialDecay(
@@ -383,6 +445,7 @@ class RandLANet(BaseModel):
 
     def get_batch_gen(self, dataset, steps_per_epoch=None, batch_size=1):
         cfg = self.cfg
+        rng = self.rng
 
         def gen():
             n_iters = dataset.num_pc if steps_per_epoch is None else steps_per_epoch * batch_size
@@ -400,26 +463,42 @@ class RandLANet(BaseModel):
                                                    pick_idx,
                                                    self.cfg.num_points)
 
-                t_normalize = cfg.get('t_normalize', {})
-                pc, feat = trans_normalize(pc, feat, t_normalize)
+                augment_cfg = self.cfg.get('augment', {}).copy()
+                val_augment_cfg = {}
+                if 'recenter' in augment_cfg:
+                    val_augment_cfg['recenter'] = augment_cfg['recenter']
+                    augment_cfg.pop('recenter')
+                if 'normalize' in augment_cfg:
+                    val_augment_cfg['normalize'] = augment_cfg['normalize']
+                    augment_cfg.pop('normalize')
+
+                self.augmenter.augment(pc,
+                                       feat,
+                                       label,
+                                       val_augment_cfg,
+                                       seed=rng)
 
                 if attr['split'] in ['training', 'train']:
-                    t_augment = cfg.get('t_augment', None)
-                    pc = trans_augment(pc, t_augment)
+                    pc, feat, label = self.augmenter.augment(pc,
+                                                             feat,
+                                                             label,
+                                                             augment_cfg,
+                                                             seed=rng)
 
                 if feat is None:
                     feat = pc.copy()
                 else:
                     feat = np.concatenate([pc, feat], axis=1)
-                assert self.cfg.dim_input == feat.shape[
-                    1], "Wrong feature dimension, please update dim_input(3 + feature_dimension) in config"
+
+                assert cfg.in_channels == feat.shape[
+                    1], "Wrong feature dimension, please update in_channels(3 + feature_dimension) in config"
 
                 yield (pc.astype(np.float32), feat.astype(np.float32),
                        label.astype(np.float32))
 
         gen_func = gen
         gen_types = (tf.float32, tf.float32, tf.int32)
-        gen_shapes = ([None, 3], [None, cfg.dim_input], [None])
+        gen_shapes = ([None, 3], [None, cfg.in_channels], [None])
 
         return gen_func, gen_types, gen_shapes
 
@@ -452,8 +531,8 @@ class RandLANet(BaseModel):
         else:
             feat = np.concatenate([pc, feat], axis=1)
 
-        assert self.cfg.dim_input == feat.shape[
-            1], "Wrong feature dimension, please update dim_input(3 + feature_dimension) in config"
+        assert self.cfg.in_channels == feat.shape[
+            1], "Wrong feature dimension, please update in_channels(3 + feature_dimension) in config"
 
         features = feat
         input_points = []
@@ -462,7 +541,7 @@ class RandLANet(BaseModel):
         input_up_samples = []
 
         for i in range(cfg.num_layers):
-            neighbour_idx = DataProcessing.knn_search(pc, pc, cfg.k_n)
+            neighbour_idx = DataProcessing.knn_search(pc, pc, cfg.num_neighbors)
 
             sub_points = pc[:pc.shape[0] // cfg.sub_sampling_ratio[i], :]
             pool_i = neighbour_idx[:pc.shape[0] // cfg.sub_sampling_ratio[i], :]
@@ -486,6 +565,9 @@ class RandLANet(BaseModel):
     def transform(self, pc, feat, label):
         cfg = self.cfg
 
+        if feat is pc:
+            feat = None
+
         input_points = []
         input_neighbors = []
         input_pools = []
@@ -493,7 +575,8 @@ class RandLANet(BaseModel):
 
         for i in range(cfg.num_layers):
             neighbour_idx = tf.numpy_function(DataProcessing.knn_search,
-                                              [pc, pc, cfg.k_n], tf.int32)
+                                              [pc, pc, cfg.num_neighbors],
+                                              tf.int32)
 
             sub_points = pc[:tf.shape(pc)[0] // cfg.sub_sampling_ratio[i], :]
             pool_i = neighbour_idx[:tf.shape(pc)[0] //
@@ -510,6 +593,11 @@ class RandLANet(BaseModel):
         input_list += [feat, label]
 
         return input_list
+
+        # input_list = input_points + input_neighbors + input_pools + input_up_samples
+        # input_list += [feat, label]
+
+        # return input_list
 
     def inference_begin(self, data):
         self.test_smooth = 0.95
@@ -570,26 +658,22 @@ class RandLANet(BaseModel):
     def preprocess(self, data, attr):
         cfg = self.cfg
 
-        points = data['point'][:, 0:3]
+        points = np.array(data['point'][:, 0:3], dtype=np.float32)
 
-        if 'label' not in data.keys() or data['label'] is None:
+        if 'label' not in data or data['label'] is None:
             labels = np.zeros((points.shape[0],), dtype=np.int32)
         else:
             labels = np.array(data['label'], dtype=np.int32).reshape((-1,))
 
-        if 'feat' not in data.keys() or data['feat'] is None:
+        if 'feat' not in data or data['feat'] is None:
             feat = None
         else:
             feat = np.array(data['feat'], dtype=np.float32)
 
         split = attr['split']
         data = dict()
-        if cfg.get('t_align', False):
-            points_min = np.expand_dims(points.min(0), 0)
-            points_min[0, :2] = 0
-            points = points - points_min
 
-        if (feat is None):
+        if feat is None:
             sub_points, sub_labels = DataProcessing.grid_subsampling(
                 points, labels=labels, grid_size=cfg.grid_size)
             sub_feat = None
