+++ ml3d/torch/models/randlanet.py
@@ -1,36 +1,15 @@
-# ----------------------------------------------------------------------------
-# -              Creative Commons BY-NC-SA 4.0 International                 -
-# ----------------------------------------------------------------------------
-
-# By exercising the Licensed Rights (defined below), You accept and agree to
-# be bound by the terms and conditions of this Creative
-# Commons Attribution-NonCommercial-ShareAlike 4.0 International Public
-# License ("Public License"). To the extent this Public License may be
-# interpreted as a contract, You are granted the Licensed Rights in
-# consideration of Your acceptance of these terms and conditions, and the
-# Licensor grants You such rights in consideration of benefits the Licensor
-# receives from making the Licensed Material available under these terms and
-# conditions.
-#
-# You can read the full license here
-# https://creativecommons.org/licenses/by-nc-sa/4.0/legalcode
-# ----------------------------------------------------------------------------
-
 import torch
 import torch.nn as nn
 import numpy as np
-import time
-from tqdm import tqdm
 
+from tqdm import tqdm
 from pathlib import Path
 from sklearn.neighbors import KDTree
-from torch.utils.tensorboard import SummaryWriter
-from torch.utils.data import Dataset, IterableDataset, DataLoader, Sampler, BatchSampler
 
-# use relative import for being compatible with Open3d main repo
 from .base_model import BaseModel
 from ..utils import helper_torch
 from ..dataloaders import DefaultBatcher
+from ...datasets.augment import SemsegAugmentation
 from ..modules.losses import filter_valid_label
 from ...datasets.utils import (DataProcessing, trans_normalize, trans_augment,
                                trans_crop_pc)
@@ -41,89 +20,319 @@ class RandLANet(BaseModel):
     """Class defining RandLANet, a Semantic Segmentation model.
     Based on the architecture
     https://arxiv.org/abs/1911.11236#
-
-    Reference Implementation - https://github.com/QingyongHu/RandLA-Net
     """
 
     def __init__(
             self,
             name='RandLANet',
-            k_n=16,  # KNN,
-            num_layers=4,  # Number of layers
-            num_points=4096 * 11,  # Number of input points
-            num_classes=19,  # Number of valid classes
+            num_neighbors=16,
+            num_layers=4,
+            num_points=4096 * 11,
+            num_classes=19,
             ignored_label_inds=[0],
             sub_sampling_ratio=[4, 4, 4, 4],
-            dim_input=3,  # 3 + feature_dimension.
-            dim_feature=8,
+            in_channels=3,  # 3 + feature_dimension.
+            dim_features=8,
             dim_output=[16, 64, 128, 256],
             grid_size=0.06,
             batcher='DefaultBatcher',
             ckpt_path=None,
             weight_decay=0.0,
+            augment={},
             **kwargs):
 
         super().__init__(name=name,
-                         k_n=k_n,
+                         num_neighbors=num_neighbors,
                          num_layers=num_layers,
                          num_points=num_points,
                          num_classes=num_classes,
                          ignored_label_inds=ignored_label_inds,
                          sub_sampling_ratio=sub_sampling_ratio,
-                         dim_input=dim_input,
-                         dim_feature=dim_feature,
+                         in_channels=in_channels,
+                         dim_features=dim_features,
                          dim_output=dim_output,
                          grid_size=grid_size,
                          batcher=batcher,
                          ckpt_path=ckpt_path,
                          weight_decay=weight_decay,
+                         augment=augment,
                          **kwargs)
         cfg = self.cfg
+        self.augmenter = SemsegAugmentation(cfg.augment, seed=self.rng)
 
-        dim_feature = cfg.dim_feature
-        self.fc0 = nn.Linear(cfg.dim_input, dim_feature)
-        self.batch_normalization = nn.BatchNorm2d(dim_feature,
-                                                  eps=1e-6,
-                                                  momentum=0.01)
-
-        d_encoder_list = []
+        self.fc0 = nn.Linear(cfg.in_channels, cfg.dim_features)
+        self.bn0 = nn.BatchNorm2d(cfg.dim_features, eps=1e-6, momentum=0.01)
 
         # Encoder
+        self.encoder = []
+        encoder_dim_list = []
+        dim_feature = cfg.dim_features
         for i in range(cfg.num_layers):
-            name = 'Encoder_layer_' + str(i)
-            self.init_dilated_res_block(dim_feature, cfg.dim_output[i], name)
-            dim_feature = cfg.dim_output[i] * 2
+            self.encoder.append(
+                LocalFeatureAggregation(dim_feature, cfg.dim_output[i],
+                                        cfg.num_neighbors))
+            dim_feature = 2 * cfg.dim_output[i]
             if i == 0:
-                d_encoder_list.append(dim_feature)
-            d_encoder_list.append(dim_feature)
+                encoder_dim_list.append(dim_feature)
+            encoder_dim_list.append(dim_feature)
 
-        feature = helper_torch.conv2d(True, dim_feature, dim_feature)
-        setattr(self, 'decoder_0', feature)
+        self.encoder = nn.ModuleList(self.encoder)
+
+        self.mlp = SharedMLP(dim_feature,
+                             dim_feature,
+                             activation_fn=nn.LeakyReLU(0.2))
 
         # Decoder
-        for j in range(cfg.num_layers):
-            name = 'Decoder_layer_' + str(j)
-            dim_input = d_encoder_list[-j - 2] + dim_feature
-            dim_output = d_encoder_list[-j - 2]
+        self.decoder = []
+        for i in range(cfg.num_layers):
+            self.decoder.append(
+                SharedMLP(encoder_dim_list[-i - 2] + dim_feature,
+                          encoder_dim_list[-i - 2],
+                          transpose=True,
+                          activation_fn=nn.LeakyReLU(0.2)))
+            dim_feature = encoder_dim_list[-i - 2]
 
-            f_decoder_i = helper_torch.conv2d_transpose(True, dim_input,
-                                                        dim_output)
-            setattr(self, name, f_decoder_i)
-            dim_feature = d_encoder_list[-j - 2]
+        self.decoder = nn.ModuleList(self.decoder)
 
-        f_layer_fc1 = helper_torch.conv2d(True, dim_feature, 64)
-        setattr(self, 'fc1', f_layer_fc1)
+        self.fc1 = nn.Sequential(
+            SharedMLP(dim_feature, 64, activation_fn=nn.LeakyReLU(0.2)),
+            SharedMLP(64, 32, activation_fn=nn.LeakyReLU(0.2)), nn.Dropout(0.5),
+            SharedMLP(32, cfg.num_classes, bn=False))
 
-        f_layer_fc2 = helper_torch.conv2d(True, 64, 32)
-        setattr(self, 'fc2', f_layer_fc2)
+    def preprocess(self, data, attr):
+        cfg = self.cfg
 
-        f_layer_fc3 = helper_torch.conv2d(False,
-                                          32,
-                                          cfg.num_classes,
-                                          activation=False)
-        setattr(self, 'fc', f_layer_fc3)
+        points = np.array(data['point'][:, 0:3], dtype=np.float32)
 
-        self.m_dropout = nn.Dropout(0.5)
+        if 'label' not in data or data['label'] is None:
+            labels = np.zeros((points.shape[0],), dtype=np.int32)
+        else:
+            labels = np.array(data['label'], dtype=np.int32).reshape((-1,))
+
+        if 'feat' not in data or data['feat'] is None:
+            feat = None
+        else:
+            feat = np.array(data['feat'], dtype=np.float32)
+
+        split = attr['split']
+        data = dict()
+
+        if feat is None:
+            sub_points, sub_labels = DataProcessing.grid_subsampling(
+                points, labels=labels, grid_size=cfg.grid_size)
+            sub_feat = None
+        else:
+            sub_points, sub_feat, sub_labels = DataProcessing.grid_subsampling(
+                points, features=feat, labels=labels, grid_size=cfg.grid_size)
+
+        search_tree = KDTree(sub_points)
+
+        data['point'] = sub_points
+        data['feat'] = sub_feat
+        data['label'] = sub_labels
+        data['search_tree'] = search_tree
+
+        if split in ["test", "testing"]:
+            proj_inds = np.squeeze(
+                search_tree.query(points, return_distance=False))
+            proj_inds = proj_inds.astype(np.int32)
+            data['proj_inds'] = proj_inds
+
+        return data
+
+    def transform(self, data, attr, min_possibility_idx=None):
+        # If num_workers > 0, use new RNG with unique seed for each thread.
+        # Else, use default RNG.
+        if torch.utils.data.get_worker_info():
+            seedseq = np.random.SeedSequence(
+                torch.utils.data.get_worker_info().seed +
+                torch.utils.data.get_worker_info().id)
+            rng = np.random.default_rng(seedseq.spawn(1)[0])
+        else:
+            rng = self.rng
+
+        cfg = self.cfg
+        inputs = dict()
+
+        pc = data['point'].copy()
+        label = data['label'].copy()
+        feat = data['feat'].copy() if data['feat'] is not None else None
+        tree = data['search_tree']
+
+        pc, selected_idxs, center_point = self.trans_point_sampler(
+            pc=pc,
+            feat=feat,
+            label=label,
+            search_tree=tree,
+            num_points=self.cfg.num_points)
+
+        label = label[selected_idxs]
+
+        if feat is not None:
+            feat = feat[selected_idxs]
+
+        augment_cfg = self.cfg.get('augment', {}).copy()
+        val_augment_cfg = {}
+        if 'recenter' in augment_cfg:
+            val_augment_cfg['recenter'] = augment_cfg['recenter']
+            augment_cfg.pop('recenter')
+        if 'normalize' in augment_cfg:
+            val_augment_cfg['normalize'] = augment_cfg['normalize']
+            augment_cfg.pop('normalize')
+
+        self.augmenter.augment(pc, feat, label, val_augment_cfg, seed=rng)
+
+        if attr['split'] in ['training', 'train']:
+            pc, feat, label = self.augmenter.augment(pc,
+                                                     feat,
+                                                     label,
+                                                     augment_cfg,
+                                                     seed=rng)
+
+        if feat is None:
+            feat = pc.copy()
+        else:
+            feat = np.concatenate([pc, feat], axis=1)
+
+        assert cfg.in_channels == feat.shape[
+            1], "Wrong feature dimension, please update dim_input(3 + feature_dimension) in config"
+
+        input_points = []
+        input_neighbors = []
+        input_pools = []
+        input_up_samples = []
+
+        for i in range(cfg.num_layers):
+            # TODO: Replace with Open3D KNN
+            neighbour_idx = DataProcessing.knn_search(pc, pc, cfg.num_neighbors)
+
+            sub_points = pc[:pc.shape[0] // cfg.sub_sampling_ratio[i], :]
+            pool_i = neighbour_idx[:pc.shape[0] // cfg.sub_sampling_ratio[i], :]
+            up_i = DataProcessing.knn_search(sub_points, pc, 1)
+            input_points.append(pc)
+            input_neighbors.append(neighbour_idx.astype(np.int64))
+            input_pools.append(pool_i.astype(np.int64))
+            input_up_samples.append(up_i.astype(np.int64))
+            pc = sub_points
+
+        inputs['coords'] = input_points
+        inputs['neighbor_indices'] = input_neighbors
+        inputs['sub_idx'] = input_pools
+        inputs['interp_idx'] = input_up_samples
+        inputs['features'] = feat
+        inputs['point_inds'] = selected_idxs
+        inputs['labels'] = label.astype(np.int64)
+
+        return inputs
+
+    def forward(self, inputs):
+        """Forward pass for RandLANet
+
+        Args:
+            inputs: torch.Tensor, shape (B, N, d_in)
+                input points
+
+        Returns
+            torch.Tensor, shape (B, num_classes, N)
+                segmentation scores for each point
+
+        """
+        cfg = self.cfg
+        feat = inputs['features'].to(self.device)  # (B, N, in_channels)
+        coords_list = [arr.to(self.device) for arr in inputs['coords']]
+        neighbor_indices_list = [
+            arr.to(self.device) for arr in inputs['neighbor_indices']
+        ]
+        subsample_indices_list = [
+            arr.to(self.device) for arr in inputs['sub_idx']
+        ]
+        interpolation_indices_list = [
+            arr.to(self.device) for arr in inputs['interp_idx']
+        ]
+
+        feat = self.fc0(feat).transpose(-2, -1).unsqueeze(
+            -1)  # (B, dim_feature, N, 1)
+        feat = self.bn0(feat)  # (B, d, N, 1)
+
+        l_relu = nn.LeakyReLU(0.2)
+        feat = l_relu(feat)
+
+        # Encoder
+        encoder_feat_list = []
+        for i in range(cfg.num_layers):
+            feat_encoder_i = self.encoder[i](coords_list[i], feat,
+                                             neighbor_indices_list[i])
+            feat_sampled_i = self.random_sample(feat_encoder_i,
+                                                subsample_indices_list[i])
+            if i == 0:
+                encoder_feat_list.append(feat_encoder_i.clone())
+            encoder_feat_list.append(feat_sampled_i.clone())
+            feat = feat_sampled_i
+
+        feat = self.mlp(feat)
+
+        # Decoder
+        for i in range(cfg.num_layers):
+            feat_interpolation_i = self.nearest_interpolation(
+                feat, interpolation_indices_list[-i - 1])
+            feat_decoder_i = torch.cat(
+                [encoder_feat_list[-i - 2], feat_interpolation_i], dim=1)
+            feat_decoder_i = self.decoder[i](feat_decoder_i)
+            feat = feat_decoder_i
+
+        scores = self.fc1(feat)
+
+        return scores.squeeze(3).transpose(1, 2)
+
+    @staticmethod
+    def random_sample(feature, pool_idx):
+        """
+        Args:
+            feature: [B, d, N, 1] input features matrix
+            pool_idx: [B, N', max_num] N' < N, N' is the selected position after pooling
+
+        Returns:
+             pool_features = [B, N', d] pooled features matrix
+        """
+        feature = feature.squeeze(3)
+        num_neigh = pool_idx.size()[2]
+        batch_size = feature.size()[0]
+        d = feature.size()[1]
+
+        pool_idx = torch.reshape(pool_idx, (batch_size, -1))
+
+        pool_idx = pool_idx.unsqueeze(2).expand(batch_size, -1, d)
+
+        feature = feature.transpose(1, 2)
+        pool_features = torch.gather(feature, 1, pool_idx)
+        pool_features = torch.reshape(pool_features,
+                                      (batch_size, -1, num_neigh, d))
+        pool_features, _ = torch.max(pool_features, 2, keepdim=True)
+        pool_features = pool_features.permute(0, 3, 1, 2)
+
+        return pool_features
+
+    @staticmethod
+    def nearest_interpolation(feature, interp_idx):
+        """
+        Args:
+            feature: [B, d, N] input features matrix
+            interp_idx: [B, up_num_points, 1] nearest neighbour index
+
+        Returns:
+             [B, up_num_points, d] interpolated features matrix
+        """
+        feature = feature.squeeze(3)
+        d = feature.size(1)
+        batch_size = interp_idx.size()[0]
+        up_num_points = interp_idx.size()[1]
+
+        interp_idx = torch.reshape(interp_idx, (batch_size, up_num_points))
+        interp_idx = interp_idx.unsqueeze(1).expand(batch_size, d, -1)
+
+        interpolatedim_features = torch.gather(feature, 2, interp_idx)
+        interpolatedim_features = interpolatedim_features.unsqueeze(3)
+        return interpolatedim_features
 
     def get_optimizer(self, cfg_pipeline):
         optimizer = torch.optim.Adam(self.parameters(),
@@ -152,87 +361,6 @@ class RandLANet(BaseModel):
         loss = Loss.weighted_CrossEntropyLoss(scores, labels)
 
         return loss, labels, scores
-
-    def update_probs(self, inputs, results, test_probs, test_labels):
-        self.test_smooth = 0.95
-
-        for b in range(results.size()[0]):
-
-            result = torch.reshape(results[b], (-1, self.cfg.num_classes))
-            probs = torch.nn.functional.softmax(result, dim=-1)
-            probs = probs.cpu().data.numpy()
-            labels = np.argmax(probs, 1)
-            inds = inputs['data']['point_inds'][b]
-
-            test_probs[inds] = self.test_smooth * test_probs[inds] + (
-                1 - self.test_smooth) * probs
-            test_labels[inds] = labels
-
-        return test_probs, test_labels
-
-    def transform(self, data, attr, min_possibility_idx=None):
-        cfg = self.cfg
-        inputs = dict()
-
-        pc = data['point'].copy()
-        label = data['label'].copy()
-        feat = data['feat'].copy() if data['feat'] is not None else None
-        tree = data['search_tree']
-
-        pc, selected_idxs, center_point = self.trans_point_sampler(
-            pc=pc,
-            feat=feat,
-            label=label,
-            search_tree=tree,
-            num_points=self.cfg.num_points)
-
-        label = label[selected_idxs]
-
-        if feat is not None:
-            feat = feat[selected_idxs]
-
-        t_normalize = cfg.get('t_normalize', {})
-        pc, feat = trans_normalize(pc, feat, t_normalize)
-
-        if attr['split'] in ['training', 'train']:
-            t_augment = cfg.get('t_augment', None)
-            pc = trans_augment(pc, t_augment)
-
-        if feat is None:
-            feat = pc.copy()
-        else:
-            feat = np.concatenate([pc, feat], axis=1)
-
-        assert cfg.dim_input == feat.shape[
-            1], "Wrong feature dimension, please update dim_input(3 + feature_dimension) in config"
-
-        features = feat
-        input_points = []
-        input_neighbors = []
-        input_pools = []
-        input_up_samples = []
-
-        for i in range(cfg.num_layers):
-            neighbour_idx = DataProcessing.knn_search(pc, pc, cfg.k_n)
-
-            sub_points = pc[:pc.shape[0] // cfg.sub_sampling_ratio[i], :]
-            pool_i = neighbour_idx[:pc.shape[0] // cfg.sub_sampling_ratio[i], :]
-            up_i = DataProcessing.knn_search(sub_points, pc, 1)
-            input_points.append(pc)
-            input_neighbors.append(neighbour_idx.astype(np.int64))
-            input_pools.append(pool_i.astype(np.int64))
-            input_up_samples.append(up_i.astype(np.int64))
-            pc = sub_points
-
-        inputs['xyz'] = input_points
-        inputs['neigh_idx'] = input_neighbors
-        inputs['sub_idx'] = input_pools
-        inputs['interp_idx'] = input_up_samples
-        inputs['features'] = features
-        inputs['point_inds'] = selected_idxs
-
-        inputs['labels'] = label.astype(np.int64)
-        return inputs
 
     def inference_begin(self, data):
         self.test_smooth = 0.95
@@ -293,298 +421,232 @@ class RandLANet(BaseModel):
         else:
             return False
 
-    def preprocess(self, data, attr):
-        cfg = self.cfg
+    def update_probs(self, inputs, results, test_probs, test_labels):
+        self.test_smooth = 0.95
 
-        points = np.array(data['point'][:, 0:3], dtype=np.float32)
+        for b in range(results.size()[0]):
 
-        if 'label' not in data.keys() or data['label'] is None:
-            labels = np.zeros((points.shape[0],), dtype=np.int32)
-        else:
-            labels = np.array(data['label'], dtype=np.int32).reshape((-1,))
+            result = torch.reshape(results[b], (-1, self.cfg.num_classes))
+            probs = torch.nn.functional.softmax(result, dim=-1)
+            probs = probs.cpu().data.numpy()
+            labels = np.argmax(probs, 1)
+            inds = inputs['data']['point_inds'][b]
 
-        if 'feat' not in data.keys() or data['feat'] is None:
-            feat = None
-        else:
-            feat = np.array(data['feat'], dtype=np.float32)
+            test_probs[inds] = self.test_smooth * test_probs[inds] + (
+                1 - self.test_smooth) * probs
+            test_labels[inds] = labels
 
-        if cfg.get('t_align', False):
-            points_min = np.expand_dims(points.min(0), 0)
-            points_min[0, :2] = 0
-            points = points - points_min
-
-        split = attr['split']
-        data = dict()
-
-        if (feat is None):
-            sub_points, sub_labels = DataProcessing.grid_subsampling(
-                points, labels=labels, grid_size=cfg.grid_size)
-            sub_feat = None
-        else:
-            sub_points, sub_feat, sub_labels = DataProcessing.grid_subsampling(
-                points, features=feat, labels=labels, grid_size=cfg.grid_size)
-
-        search_tree = KDTree(sub_points)
-
-        data['point'] = sub_points
-        data['feat'] = sub_feat
-        data['label'] = sub_labels
-        data['search_tree'] = search_tree
-
-        if split in ["test", "testing"]:
-            proj_inds = np.squeeze(
-                search_tree.query(points, return_distance=False))
-            proj_inds = proj_inds.astype(np.int32)
-            data['proj_inds'] = proj_inds
-
-        return data
-
-    def init_att_pooling(self, d, dim_output, name):
-        att_activation = nn.Linear(d, d)
-        setattr(self, name + 'fc', att_activation)
-
-        f_agg = helper_torch.conv2d(True, d, dim_output)
-        setattr(self, name + 'mlp', f_agg)
-
-    def init_building_block(self, dim_input, dim_output, name):
-        f_pc = helper_torch.conv2d(True, 10, dim_input)
-        setattr(self, name + 'mlp1', f_pc)
-
-        self.init_att_pooling(dim_input * 2, dim_output // 2,
-                              name + 'att_pooling_1')
-
-        f_xyz = helper_torch.conv2d(True, dim_input, dim_output // 2)
-        setattr(self, name + 'mlp2', f_xyz)
-
-        self.init_att_pooling(dim_input * 2, dim_output, name + 'att_pooling_2')
-
-    def init_dilated_res_block(self, dim_input, dim_output, name):
-        f_pc = helper_torch.conv2d(True, dim_input, dim_output // 2)
-        setattr(self, name + 'mlp1', f_pc)
-
-        self.init_building_block(dim_output // 2, dim_output, name + 'LFA')
-
-        f_pc = helper_torch.conv2d(True,
-                                   dim_output,
-                                   dim_output * 2,
-                                   activation=False)
-        setattr(self, name + 'mlp2', f_pc)
-
-        shortcut = helper_torch.conv2d(True,
-                                       dim_input,
-                                       dim_output * 2,
-                                       activation=False)
-        setattr(self, name + 'shortcut', shortcut)
-
-    def forward_gather_neighbour(self, pc, neighbor_idx):
-        # pc:           BxNxd
-        # neighbor_idx: BxNxK
-        B, N, K = neighbor_idx.size()
-        d = pc.size()[2]
-
-        extended_idx = neighbor_idx.unsqueeze(1).expand(B, d, N, K)
-        extended_coords = pc.transpose(-2, -1).unsqueeze(-1).expand(B, d, N, K)
-        features = torch.gather(extended_coords, 2, extended_idx)
-
-        return features
-
-    def forward_att_pooling(self, feature_set, name):
-        # feature_set: BxdxNxK
-        batch_size = feature_set.size()[0]
-        num_points = feature_set.size()[2]
-        num_neigh = feature_set.size()[3]
-        d = feature_set.size()[1]
-
-        f_reshaped = torch.reshape(feature_set.permute(0, 2, 3, 1),
-                                   (-1, num_neigh, d))
-
-        m_dense = getattr(self, name + 'fc')
-        att_activation = m_dense(f_reshaped)
-
-        m_softmax = nn.Softmax(dim=1)
-        att_scores = m_softmax(att_activation)
-
-        # print("att_scores = ", att_scores.shape)
-        f_agg = f_reshaped * att_scores
-        f_agg = torch.sum(f_agg, dim=1, keepdim=True)
-        f_agg = torch.reshape(f_agg, (batch_size, num_points, 1, d))
-        f_agg = f_agg.permute(0, 3, 1, 2)
-
-        m_conv2d = getattr(self, name + 'mlp')
-        f_agg = m_conv2d(f_agg)
-
-        return f_agg
-
-    def forward_relative_pos_encoding(self, xyz, neigh_idx):
-        B, N, K = neigh_idx.size()
-        neighbor_xyz = self.forward_gather_neighbour(xyz, neigh_idx)
-
-        xyz_tile = xyz.transpose(-2, -1).unsqueeze(-1).expand(B, 3, N, K)
-        #xyz_tile = xyz.unsqueeze(2).repeat(1, 1, neigh_idx.size()[-1], 1)
-
-        relative_xyz = xyz_tile - neighbor_xyz
-        relative_dis = torch.sqrt(
-            torch.sum(torch.square(relative_xyz), dim=1, keepdim=True))
-        relative_feature = torch.cat(
-            [relative_dis, relative_xyz, xyz_tile, neighbor_xyz], axis=1)
-
-        return relative_feature
-
-    def forward_building_block(self, xyz, feature, neigh_idx, name):
-        f_xyz = self.forward_relative_pos_encoding(xyz, neigh_idx)
-        m_conv2d = getattr(self, name + 'mlp1')
-        f_xyz = m_conv2d(f_xyz)
-
-        feature = feature.transpose(1, 2)
-        f_neighbours = self.forward_gather_neighbour(
-            torch.squeeze(feature, axis=3), neigh_idx)
-        f_concat = torch.cat([f_neighbours, f_xyz], axis=1)
-
-        f_pc_agg = self.forward_att_pooling(f_concat, name + 'att_pooling_1')
-
-        m_conv2d = getattr(self, name + 'mlp2')
-        f_xyz = m_conv2d(f_xyz)
-
-        f_pc_agg = f_pc_agg.transpose(1, 2)
-        f_neighbours = self.forward_gather_neighbour(
-            torch.squeeze(f_pc_agg, axis=3), neigh_idx)
-        f_concat = torch.cat([f_neighbours, f_xyz], axis=1)
-        f_pc_agg = self.forward_att_pooling(f_concat, name + 'att_pooling_2')
-
-        return f_pc_agg
-
-    def forward_dilated_res_block(self, feature, xyz, neigh_idx, dim_output,
-                                  name):
-        m_conv2d = getattr(self, name + 'mlp1')
-        f_pc = m_conv2d(feature)
-
-        f_pc = self.forward_building_block(xyz, f_pc, neigh_idx, name + 'LFA')
-
-        m_conv2d = getattr(self, name + 'mlp2')
-        f_pc = m_conv2d(f_pc)
-
-        m_conv2d = getattr(self, name + 'shortcut')
-        shortcut = m_conv2d(feature)
-
-        m_leakyrelu = nn.LeakyReLU(0.2)
-
-        result = m_leakyrelu(f_pc + shortcut)
-        return result
-
-    def forward(self, inputs):
-        device = self.device
-        xyz = [arr.to(device) for arr in inputs['xyz']]
-        neigh_idx = [arr.to(device) for arr in inputs['neigh_idx']]
-        interp_idx = [arr.to(device) for arr in inputs['interp_idx']]
-        sub_idx = [arr.to(device) for arr in inputs['sub_idx']]
-        feature = inputs['features'].to(device)
-
-        m_dense = getattr(self, 'fc0')
-        feature = m_dense(feature).transpose(-2, -1).unsqueeze(-1)
-
-        m_bn = getattr(self, 'batch_normalization')
-        feature = m_bn(feature)
-
-        m_leakyrelu = nn.LeakyReLU(0.2)
-        feature = m_leakyrelu(feature)
-
-        # B d N 1
-        # B N 1 d
-        # Encoder
-        f_encoder_list = []
-        for i in range(self.cfg.num_layers):
-            name = 'Encoder_layer_' + str(i)
-            f_encoder_i = self.forward_dilated_res_block(
-                feature, xyz[i], neigh_idx[i], self.cfg.dim_output[i], name)
-            f_sampled_i = self.random_sample(f_encoder_i, sub_idx[i])
-            feature = f_sampled_i
-            if i == 0:
-                f_encoder_list.append(f_encoder_i)
-            f_encoder_list.append(f_sampled_i)
-
-        m_conv2d = getattr(self, 'decoder_0')
-        feature = m_conv2d(f_encoder_list[-1])
-
-        # Decoder
-        f_decoder_list = []
-        for j in range(self.cfg.num_layers):
-            f_interp_i = self.nearest_interpolation(feature, interp_idx[-j - 1])
-            name = 'Decoder_layer_' + str(j)
-
-            m_transposeconv2d = getattr(self, name)
-            concat_feature = torch.cat([f_encoder_list[-j - 2], f_interp_i],
-                                       dim=1)
-            f_decoder_i = m_transposeconv2d(concat_feature)
-
-            feature = f_decoder_i
-            f_decoder_list.append(f_decoder_i)
-
-        m_conv2d = getattr(self, 'fc1')
-        f_layer_fc1 = m_conv2d(f_decoder_list[-1])
-
-        m_conv2d = getattr(self, 'fc2')
-        f_layer_fc2 = m_conv2d(f_layer_fc1)
-
-        f_layer_drop = self.m_dropout(f_layer_fc2)
-
-        test_hidden = f_layer_fc2.permute(0, 2, 3, 1)
-
-        m_conv2d = getattr(self, 'fc')
-        f_layer_fc3 = m_conv2d(f_layer_drop)
-
-        f_out = f_layer_fc3.squeeze(3).transpose(1, 2)
-
-        return f_out
-
-    @staticmethod
-    def random_sample(feature, pool_idx):
-        """
-        Args:
-            feature: [B, d, N, 1] input features matrix
-            pool_idx: [B, N', max_num] N' < N, N' is the selected position after pooling
-
-        Returns:
-             pool_features = [B, N', d] pooled features matrix
-        """
-        feature = feature.squeeze(3)
-        num_neigh = pool_idx.size()[2]
-        batch_size = feature.size()[0]
-        d = feature.size()[1]
-
-        pool_idx = torch.reshape(pool_idx, (batch_size, -1))
-
-        pool_idx = pool_idx.unsqueeze(2).expand(batch_size, -1, d)
-
-        feature = feature.transpose(1, 2)
-        pool_features = torch.gather(feature, 1, pool_idx)
-        pool_features = torch.reshape(pool_features,
-                                      (batch_size, -1, num_neigh, d))
-        pool_features, _ = torch.max(pool_features, 2, keepdim=True)
-        pool_features = pool_features.permute(0, 3, 1, 2)
-
-        return pool_features
-
-    @staticmethod
-    def nearest_interpolation(feature, interp_idx):
-        """
-        Args:
-            feature: [B, d, N] input features matrix
-            interp_idx: [B, up_num_points, 1] nearest neighbour index
-
-        Returns:
-             [B, up_num_points, d] interpolated features matrix
-        """
-        feature = feature.squeeze(3)
-        d = feature.size(1)
-        batch_size = interp_idx.size()[0]
-        up_num_points = interp_idx.size()[1]
-
-        interp_idx = torch.reshape(interp_idx, (batch_size, up_num_points))
-        interp_idx = interp_idx.unsqueeze(1).expand(batch_size, d, -1)
-
-        interpolatedim_features = torch.gather(feature, 2, interp_idx)
-        interpolatedim_features = interpolatedim_features.unsqueeze(3)
-        return interpolatedim_features
+        return test_probs, test_labels
 
 
 MODEL._register_module(RandLANet, 'torch')
+
+
+class SharedMLP(nn.Module):
+
+    def __init__(self,
+                 in_channels,
+                 out_channels,
+                 kernel_size=1,
+                 stride=1,
+                 transpose=False,
+                 bn=True,
+                 activation_fn=None):
+        super(SharedMLP, self).__init__()
+
+        if transpose:
+            self.conv = nn.ConvTranspose2d(in_channels,
+                                           out_channels,
+                                           kernel_size=kernel_size,
+                                           stride=stride,
+                                           padding=(kernel_size - 1) // 2)
+        else:
+            self.conv = nn.Conv2d(in_channels,
+                                  out_channels,
+                                  kernel_size=kernel_size,
+                                  stride=stride,
+                                  padding=(kernel_size - 1) // 2)
+
+        self.batch_norm = nn.BatchNorm2d(out_channels, eps=1e-6,
+                                         momentum=0.01) if bn else None
+        self.activation_fn = activation_fn
+
+    def forward(self, input):
+        """Forward pass of the Module.
+
+        Args:
+            input: torch.Tensor of shape (B, dim_in, N, K)
+
+        Returns:
+            torch.Tensor, shape (B, dim_out, N, K)
+        """
+        x = self.conv(input)
+        if self.batch_norm:
+            x = self.batch_norm(x)
+        if self.activation_fn:
+            x = self.activation_fn(x)
+        return x
+
+
+class LocalSpatialEncoding(nn.Module):
+
+    def __init__(self, dim_in, dim_out, num_neighbors, encode_pos=False):
+        super(LocalSpatialEncoding, self).__init__()
+
+        self.num_neighbors = num_neighbors
+        self.mlp = SharedMLP(dim_in, dim_out, activation_fn=nn.LeakyReLU(0.2))
+        self.encode_pos = encode_pos
+
+        # self.device = device
+
+    def gather_neighbor(self, coords, neighbor_indices):
+        """Gather features based on neighbor indices.
+
+        Args:
+            coords: torch.Tensor of shape (B, N, d)
+            neighbor_indices: torch.Tensor of shape (B, N, K)
+        
+        Returns:
+            gathered neighbors of shape (B, dim, N, K)
+
+        """
+        B, N, K = neighbor_indices.size()
+        dim = coords.shape[2]
+
+        extended_indices = neighbor_indices.unsqueeze(1).expand(B, dim, N, K)
+        extended_coords = coords.transpose(-2, -1).unsqueeze(-1).expand(
+            B, dim, N, K)
+        neighbor_coords = torch.gather(extended_coords, 2,
+                                       extended_indices)  # (B, dim, N, K)
+
+        return neighbor_coords
+
+    def forward(self,
+                coords,
+                features,
+                neighbor_indices,
+                relative_features=None):
+        """Forward pass of the Module.
+
+        Args:
+            coords: coordinates of the pointcloud
+                torch.Tensor of shape (B, N, 3)
+            features: features of the pointcloud.
+                torch.Tensor of shape (B, d, N, 1)
+            neighbor_indices: indices of k neighbours.
+                torch.Tensor of shape (B, N, K)
+
+        Returns:
+            torch.Tensor of shape (B, 2*d, N, K)
+        """
+        # finding neighboring points
+        B, N, K = neighbor_indices.size()
+
+        if self.encode_pos:
+            neighbor_coords = self.gather_neighbor(coords, neighbor_indices)
+
+            extended_coords = coords.transpose(-2, -1).unsqueeze(-1).expand(
+                B, 3, N, K)
+
+            relative_pos = extended_coords - neighbor_coords
+            relative_dist = torch.sqrt(
+                torch.sum(torch.square(relative_pos), dim=1, keepdim=True))
+
+            relative_features = torch.cat(
+                [relative_dist, relative_pos, extended_coords, neighbor_coords],
+                dim=1)
+
+        else:
+            if relative_features is None:
+                raise ValueError(
+                    "LocalSpatialEncoding: Require relative_features for second pass."
+                )
+
+        relative_features = self.mlp(relative_features)
+
+        neighbor_features = self.gather_neighbor(
+            features.transpose(1, 2).squeeze(3), neighbor_indices)
+
+        return torch.cat([neighbor_features, relative_features],
+                         dim=1), relative_features
+
+
+class AttentivePooling(nn.Module):
+
+    def __init__(self, in_channels, out_channels):
+        super(AttentivePooling, self).__init__()
+
+        self.score_fn = nn.Sequential(nn.Linear(in_channels, in_channels),
+                                      nn.Softmax(dim=-2))
+        self.mlp = SharedMLP(in_channels,
+                             out_channels,
+                             activation_fn=nn.LeakyReLU(0.2))
+
+    def forward(self, x):
+        """Forward pass of the Module.
+
+        Args:
+            x: torch.Tensor of shape (B, dim_in, N, K).
+
+        Returns:
+            torch.Tensor of shape (B, d_out, N, 1).
+        """
+        # computing attention scores
+        scores = self.score_fn(x.permute(0, 2, 3, 1)).permute(0, 3, 1, 2)
+
+        # sum over the neighbors
+        features = torch.sum(scores * x, dim=-1,
+                             keepdim=True)  # shape (B, d_in, N, 1)
+
+        return self.mlp(features)
+
+
+class LocalFeatureAggregation(nn.Module):
+
+    def __init__(self, d_in, d_out, num_neighbors):
+        super(LocalFeatureAggregation, self).__init__()
+
+        self.num_neighbors = num_neighbors
+
+        self.mlp1 = SharedMLP(d_in, d_out // 2, activation_fn=nn.LeakyReLU(0.2))
+        self.lse1 = LocalSpatialEncoding(10,
+                                         d_out // 2,
+                                         num_neighbors,
+                                         encode_pos=True)
+        self.pool1 = AttentivePooling(d_out, d_out // 2)
+
+        self.lse2 = LocalSpatialEncoding(d_out // 2, d_out // 2, num_neighbors)
+        self.pool2 = AttentivePooling(d_out, d_out)
+        self.mlp2 = SharedMLP(d_out, 2 * d_out)
+
+        self.shortcut = SharedMLP(d_in, 2 * d_out)
+        self.lrelu = nn.LeakyReLU()
+
+    def forward(self, coords, feat, neighbor_indices):
+        """Forward pass of the Module.
+
+        Args:
+            coords: coordinates of the pointcloud
+                torch.Tensor of shape (B, N, 3).
+            feat: features of the pointcloud.
+                torch.Tensor of shape (B, d, N, 1)
+            neighbor_indices: Indices of neighbors.
+
+        Returns:
+            torch.Tensor of shape (B, 2*d_out, N, 1).
+
+        """
+        # knn_output = knn(coords.cpu().contiguous(), coords.cpu().contiguous(), self.num_neighbors)
+
+        x = self.mlp1(feat)
+
+        x, neighbor_features = self.lse1(coords, x, neighbor_indices)
+        x = self.pool1(x)
+
+        x, _ = self.lse2(coords,
+                         x,
+                         neighbor_indices,
+                         relative_features=neighbor_features)
+        x = self.pool2(x)
+
+        return self.lrelu(self.mlp2(x) + self.shortcut(feat))
